refactor: 필터 타입/헬퍼 분리
